pyproject.toml
[project]
name = "expense-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastmcp>=2.14.1",
    "psycopg2-binary>=2.9.11",
    "starlette>=0.50.0",
]
